# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API MAK3RS DOCUMENTATION",
        "description": "",
        "contact": {
            "email": "info@mak3rs.tk"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/communities/collect": {
            "get": {
                "tags": [
                    "Collect Control"
                ],
                "description": "Obtenemos todas las recogidas",
                "operationId": "App\\Http\\Controllers\\CollectControlController::getCollectControl",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "community": {
                                        "description": "",
                                        "type": "string"
                                    },
                                    "user": {
                                        "description": "",
                                        "type": "string"
                                    },
                                    "status": {
                                        "description": "export",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": ""
                    },
                    "404": {
                        "description": ""
                    },
                    "500": {
                        "description": ""
                    }
                }
            }
        },
        "/communities/collect/add": {
            "post": {
                "tags": [
                    "Collect Control"
                ],
                "description": "Añadimos una pieza a una recogida",
                "operationId": "App\\Http\\Controllers\\CollectControlController::add",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "community": {
                                        "description": "",
                                        "type": "string"
                                    },
                                    "user": {
                                        "description": "",
                                        "type": "string"
                                    },
                                    "status": {
                                        "description": "",
                                        "type": "string"
                                    },
                                    "pieces": {
                                        "description": "",
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    },
                                    "address": {
                                        "description": "",
                                        "type": "string"
                                    },
                                    "location": {
                                        "description": "",
                                        "type": "string"
                                    },
                                    "province": {
                                        "description": "",
                                        "type": "string"
                                    },
                                    "state": {
                                        "description": "",
                                        "type": "string"
                                    },
                                    "country": {
                                        "description": "",
                                        "type": "string"
                                    },
                                    "address_description": {
                                        "description": "",
                                        "type": "string"
                                    },
                                    "cp": {
                                        "description": "",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": ""
                    },
                    "404": {
                        "description": ""
                    },
                    "500": {
                        "description": ""
                    }
                }
            }
        },
        "/communities/all": {
            "get": {
                "tags": [
                    "Community"
                ],
                "description": "Obtenemos todas las comunidades",
                "operationId": "App\\Http\\Controllers\\CommunityController::communities",
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "alias": {
                                        "description": "",
                                        "type": "string"
                                    },
                                    "name": {
                                        "description": "",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "List Communities"
                    }
                }
            }
        },
...
        "/users/communities": {
            "get": {
                "tags": [
                    "User"
                ],
                "description": "Listado de comunidades a las que pertenece el usuario",
                "operationId": "App\\Http\\Controllers\\UserController::communities",
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "uuid": {
                                        "description": "User",
                                        "type": "string"
                                    },
                                    "alias": {
                                        "description": "",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": ""
                    }
                }
            }
        }
    }
}